requirements.txt
numpy
Pillow
Flask
mediapipe
opencv-python